[GOPHERMART-1] fixed int64

+++ internal/models/orders.go
@@ -8,12 +8,12 @@ import (
 )
 
 type Order struct {
-	ID         string `json:"order_id"    db:"order_id"`
-	UserID     string `json:"user_id"     db:"user_id"`
-	Number     string `json:"number"      db:"number"`
-	Status     string `json:"status"      db:"status"`
-	Accrual    int64  `json:"accrual"     db:"accrual"`
-	UploadedAt string `json:"uploaded_at" db:"uploaded_at"`
+	ID         string  `json:"order_id"    db:"order_id"`
+	UserID     string  `json:"user_id"     db:"user_id"`
+	Number     string  `json:"number"      db:"number"`
+	Status     string  `json:"status"      db:"status"`
+	Accrual    float64 `json:"accrual"     db:"accrual"`
+	UploadedAt string  `json:"uploaded_at" db:"uploaded_at"`
 }
 
 func NewOrder(userID string, number string) *Order {
